map container removed
