+++ components/map/zoom/component.tsx
@@ -10,7 +10,7 @@ import { ViewportProps } from 'react-map-gl';
 export interface ZoomControlProps {
   viewport: Partial<ViewportProps>;
   className?: string;
-  onZoomChange: (zoom: number) => void;
+  onZoomChange: (z) => void;
 }
 
 export const ZoomControl = ({
