+++ package.json
@@ -13,32 +13,36 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.4.0",
     "autoprefixer": "^10.1.0",
-    "classnames": "^2.2.6",
-    "@zeit/next-sass": "^1.0.1",
     "axios": "^0.20.0",
     "babel-eslint": "^10.1.0",
     "babel-plugin-module-resolver": "^4.0.0",
+    "classnames": "^2.2.6",
     "d3-ease": "^2.0.0",
     "deck.gl": "7.3.6",
     "layer-manager": "^3.0.8",
     "lodash": "^4.17.20",
     "luma.gl": "7.3.2",
     "mapbox-gl": "^1.12.0",
-    "next": "^9.5.3",
+    "next": "^10.0.7",
     "next-redux-wrapper": "^6.0.2",
     "postcss": "^8.2.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-map-gl": "^6.1.8",
     "react-redux": "^7.2.1",
     "redux": "^4.0.5",
-    "use-debounce": "^5.2.0",
-    "tailwindcss": "^2.0.2"
+    "tailwindcss": "^2.0.2",
+    "use-debounce": "^5.2.0"
   },
   "devDependencies": {
+    "@math.gl/web-mercator": "^3.3.2",
     "@types/node": "^14.14.10",
     "@types/react": "^17.0.0",
     "@typescript-eslint/eslint-plugin": "^4.9.0",
     "@typescript-eslint/parser": "^4.9.0",
+    "axios": "^0.20.0",
+    "babel-eslint": "^10.1.0",
+    "babel-plugin-module-resolver": "^4.0.0",
     "cypress": "^6.1.0",
     "eslint": "7.2.0",
     "eslint-config-airbnb-typescript": "^12.0.0",
@@ -49,15 +53,11 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "husky": "^4.3.0",
     "lint-staged": "^9.2.5",
-    "stylelint-config-recommended": "^3.0.0",
-    "typescript": "^4.1.2",
-    "axios": "^0.20.0",
-    "babel-eslint": "^10.1.0",
-    "babel-plugin-module-resolver": "^4.0.0",
     "mapbox-gl": "^1.12.0",
-    "@math.gl/web-mercator": "^3.3.2",
     "next-redux-wrapper": "^6.0.2",
-    "redux-devtools-extension": "^2.13.8"
+    "redux-devtools-extension": "^2.13.8",
+    "stylelint-config-recommended": "^3.0.0",
+    "typescript": "^4.1.2"
   },
   "lint-staged": {
     "*.{js,jsx}": [
