+++ tailwind.config.js
@@ -5,11 +5,10 @@ module.exports = {
   darkMode: false,
   theme: {
     colors: {
+      ...colors,
       color1: '#026BFE',
       color2: '#FD1B30',
       color3: '#6907E8',
-      white: colors.white,
-      black: colors.black,
     },
     pseudo: {
       before: 'before',
